Remove entity scan
@SpringBootApplication enables various auto-configurations depending on the starter used.
Among them is everything you'll need for hibernate, including an entity scan down from the class
containing @SpringBootApplication, hence the structure suggested in the previous commit.

That is the official recommended way. I'd prefer as reader having it that way.

Maybe you want to add ‘@EntityScan' to the tip itself as additional hint, but not in the repo.